# paddlelabel_ml/openapi.yml
openapi: 3.0.0
info:
  title: PP Label ML backend API
  version: 0.1.0
servers:
  - url: "http://localhost:1234"
paths:
  /:
    get:
      summary: Get all models
      tags:
        - model
      responses:
        "200":
          description: All models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Model"
      x-openapi-router-controller: paddlelabel_ml.api
      operationId: getAll
      description: Get all models
  /running:
    get:
      summary: Detect ml backend running
      tags:
        - model
      responses:
        "200":
          description: ml backend running
          content:
            application/json:
              schema:
                type: string
      x-openapi-router-controller: paddlelabel_ml.api
      operationId: isBackendUp
      description: Detect ml backend running
  "/{model_name}/train":
    post:
      parameters:
        - $ref: "#/components/parameters/model_name"
      summary: Train a model
      tags:
        - model
      x-openapi-router-controller: paddlelabel_ml.api
      operationId: train
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                data_dir:
                  type: string
                configs:
                  type: object
              required:
                - data_dir
                - configs
      description: Train a model
      responses:
        "200":
          description: Train model
    get:
      parameters:
        - $ref: "#/components/parameters/model_name"
      tags:
        - model
      summary: Get training progress
      x-openapi-router-controller: paddlelabel_ml.api
      operationId: getProgress
      description: Get training progress
      responses:
        "200":
          description: Get training progress
  "/{model_name}/eval":
    post:
      parameters:
        - $ref: "#/components/parameters/model_name"
      tags:
        - model
      summary: Evaluate model performance
      x-openapi-router-controller: paddlelabel_ml.api
      operationId: evaluate
      description: Evaluate model performance
      responses:
        "200":
          description: Evaluate model performance
  "/{model_name}/predict":
    post:
      parameters:
        - $ref: "#/components/parameters/model_name"
      summary: Predict
      tags:
        - model
      x-openapi-router-controller: paddlelabel_ml.api
      operationId: predict
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                img:
                  type: string
                format:
                  type: string
                # create_label:
                #   type: boolean
                piggyback:
                  type: object
                other:
                  type: object
                  properties:
                    clicks:
                      type: array
                      items: {}
              required:
                - img
                - format
      description: Predict
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
  "/{model_name}/load":
    post:
      parameters:
        - $ref: "#/components/parameters/model_name"
      summary: Load a model into memory. Model must be loaded before being used.
      tags:
        - model
      x-openapi-router-controller: paddlelabel_ml.api
      operationId: load
      description: Load model
      responses:
        "200":
          description: Result
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                init_params:
                  type: object
  "/{model_name}/unload":
    get:
      parameters:
        - $ref: "#/components/parameters/model_name"
      summary: Unload a model to free up memory
      tags:
        - model
      x-openapi-router-controller: paddlelabel_ml.api
      operationId: unload
      description: Unload model
      responses:
        "200":
          description: Result
components:
  schemas:
    Model:
      title: Model
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        task_categories:
          type: string
        interactive:
          type: boolean
        label_names:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
  parameters:
    model_name:
      name: model_name
      in: path
      schema:
        type: string
        example: EISeg
      description: model name
      required: true